feat: add auth routes
