+++ backend/app.js
@@ -9,6 +9,7 @@ const passport = require("passport");
 const userRoutes = require("./routes/userRoutes");
 const entryRoutes = require("./routes/entryRoutes");
 const docRoutes = require("./routes/docRoutes");
+const authRoutes = require("./routes/authRoutes");
 
 require("./middleware/passport");
 
@@ -26,6 +27,7 @@ if (process.env.NODE_ENV === "development") {
 
 // API Routes
 
+app.use("/api/auth", authRoutes);
 app.use("/api/docs", docRoutes);
 app.use("/api/users", userRoutes);
 app.use(
